Create 3 tools to improve mobile coverage
If you Google “how to improve my cell phone signal”, the search engine will present you with three (3) suggestions. Adhesives amplifiers, antennas and cellular signal repeaters.

Some of these cost as little as 9 euros, while others are tens of times more expensive. But, the question that arises is: will any of these products work? Which is the best option? Below, we will review the most popular solutions to amplify the signal and answer with this study if they are worth it.